backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File Handling
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==0.8.11

# NLP and Machine Learning (using CPU-only torch for smaller size)
spacy==3.7.2
sentence-transformers==2.2.2
torch==2.0.1+cpu

# Data Processing (pre-built wheels)
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS Support for Frontend
fastapi-cors>=0.0.6

# Logging and Debugging
python-json-logger==2.0.7
